feat: move query anonymizer to engine rewrite
This is so that we can avoid a circular dependency between
parser and engine
