+++ ksqldb-engine/src/test/java/io/confluent/ksql/engine/rewrite/QueryAnonymizerTest.java
@@ -1,19 +1,16 @@
 /*
  * Copyright 2021 Confluent Inc.
- *
  * Licensed under the Confluent Community License (the "License"); you may not use
  * this file except in compliance with the License.  You may obtain a copy of the
  * License at
- *
  * http://www.confluent.io/confluent-community-license
- *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
  * WARRANTIES OF ANY KIND, either express or implied.  See the License for the
  * specific language governing permissions and limitations under the License.
  */
 
-package io.confluent.ksql.util;
+package io.confluent.ksql.engine.rewrite;
 
 import io.confluent.ksql.parser.DefaultKsqlParser;
 import io.confluent.ksql.parser.ParsingException;
